+++ magpie/api/management/__init__.py
@@ -1,9 +1,9 @@
-import logging
-logger = logging.getLogger(__name__)
+from magpie.common import get_logger
+LOGGER = get_logger(__name__)
 
 
 def includeme(config):
-    logger.info('Adding management routes ...')
+    LOGGER.info('Adding management routes ...')
     config.include('magpie.api.management.group')
     config.include('magpie.api.management.user')
     config.include('magpie.api.management.service')