+++ magpie/api/management/user/user_views.py
@@ -14,10 +14,9 @@ from magpie.definitions.pyramid_definitions import (
 )
 from magpie.definitions.ziggurat_definitions import UserService, GroupService
 from magpie.constants import get_constant
-from magpie.common import str2bool
+from magpie.common import str2bool, get_logger
 from magpie import models
-import logging
-LOGGER = logging.getLogger(__name__)
+LOGGER = get_logger(__name__)
 
 
 @s.UsersAPI.get(tags=[s.UsersTag], response_schemas=s.Users_GET_responses)
@@ -219,8 +218,8 @@ def get_user_resources_view(request):
 def get_user_inherited_resources_view(request):
     """[DEPRECATED: use '/users/{user_name}/resources?inherit=true']
     List all resources a user has permissions on with his inherited user and groups permissions."""
-    LOGGER.warn("Route deprecated: [{0}], Instead Use: [{1}]"
-                .format(s.UserInheritedResourcesAPI.path, s.UserResourcesAPI.path + "?inherit=true"))
+    LOGGER.warning("Route deprecated: [{0}], Instead Use: [{1}]"
+                   .format(s.UserInheritedResourcesAPI.path, s.UserResourcesAPI.path + "?inherit=true"))
     return HTTPMovedPermanently(location=request.path.replace('/inherited_resources', '/resources?inherit=true'))
 
 
@@ -253,9 +252,9 @@ def get_user_resource_permissions_view(request):
 def get_user_resource_inherit_groups_permissions_view(request):
     """[DEPRECATED: use '/users/{user_name}/resources/{resource_id}/permissions?inherit=true']
     List all permissions a user has on a specific resource with his inherited user and groups permissions."""
-    LOGGER.warn("Route deprecated: [{0}], Instead Use: [{1}]"
-                .format(s.UserResourceInheritedPermissionsAPI.path,
-                        s.UserResourcePermissionsAPI.path + "?inherit=true"))
+    LOGGER.warning("Route deprecated: [{0}], Instead Use: [{1}]"
+                   .format(s.UserResourceInheritedPermissionsAPI.path,
+                           s.UserResourcePermissionsAPI.path + "?inherit=true"))
     return HTTPMovedPermanently(location=request.path.replace('/inherited_permissions', '/permissions?inherit=true'))
 
 
@@ -314,8 +313,8 @@ def get_user_services_view(request):
 def get_user_inherited_services_view(request):
     """[DEPRECATED: use '/users/{user_name}/services?inherit=true']
     List all services a user has permissions on with his inherited user and groups permissions."""
-    LOGGER.warn("Route deprecated: [{0}], Instead Use: [{1}]"
-                .format(s.LoggedUserInheritedServicesAPI.path, s.LoggedUserServicesAPI.path + "?inherit=true"))
+    LOGGER.warning("Route deprecated: [{0}], Instead Use: [{1}]"
+                   .format(s.LoggedUserInheritedServicesAPI.path, s.LoggedUserServicesAPI.path + "?inherit=true"))
     return HTTPMovedPermanently(location=request.path.replace('/inherited_services', '/services?inherit=true'))
 
 
@@ -330,8 +329,9 @@ def get_user_inherited_services_view(request):
 def get_user_service_inherited_permissions_view(request):
     """[DEPRECATED: use '/users/{user_name}/services/{service_name}/permissions?inherit=true']
     List all permissions a user has on a service using all his inherited user and groups permissions."""
-    LOGGER.warn("Route deprecated: [{0}], Instead Use: [{1}]"
-                .format(s.UserServiceInheritedPermissionsAPI.path, s.UserServicePermissionsAPI.path + "?inherit=true"))
+    LOGGER.warning("Route deprecated: [{0}], Instead Use: [{1}]"
+                   .format(s.UserServiceInheritedPermissionsAPI.path,
+                           s.UserServicePermissionsAPI.path + "?inherit=true"))
     return HTTPMovedPermanently(location=request.path.replace('/inherited_permissions', '/permissions?inherit=true'))
 
 
@@ -419,6 +419,6 @@ def get_user_service_resources_view(request):
 def get_user_service_inherited_resources_view(request):
     """[DEPRECATED: use '/users/{user_name}/services/{service_name}/resources?inherit=true']
     List all resources under a service a user has permission on using all his inherited user and groups permissions."""
-    LOGGER.warn("Route deprecated: [{0}], Instead Use: [{1}]"
-                .format(s.UserServiceInheritedResourcesAPI.path, s.UserServiceResourcesAPI.path + "?inherit=true"))
+    LOGGER.warning("Route deprecated: [{0}], Instead Use: [{1}]"
+                   .format(s.UserServiceInheritedResourcesAPI.path, s.UserServiceResourcesAPI.path + "?inherit=true"))
     return HTTPMovedPermanently(location=request.path.replace('/inherited_resources', '/resources?inherit=true'))