+++ magpie/constants.py
@@ -2,10 +2,11 @@
 # -*- coding: utf-8 -*-
 import os
 import shutil
+# noinspection PyPackageRequirements
 import dotenv
 import logging
-from magpie.common import str2bool, raise_log, print_log, get_settings_from_config_ini
-logger = logging.getLogger(__name__)
+from magpie.common import str2bool, raise_log, print_log, get_settings_from_config_ini, get_logger
+LOGGER = get_logger(__name__)
 
 # ===========================
 # path variables
@@ -44,7 +45,7 @@ try:
     dotenv.load_dotenv(MAGPIE_ENV_FILE, override=False)
     dotenv.load_dotenv(MAGPIE_POSTGRES_ENV_FILE, override=False)
 except IOError:
-    logger.warn("Failed to open environment files [MAGPIE_ENV_DIR={}].".format(MAGPIE_ENV_DIR))
+    LOGGER.warning("Failed to open environment files [MAGPIE_ENV_DIR={}].".format(MAGPIE_ENV_DIR))
     pass
 
 # get default configurations from ini file