+++ magpie/definitions/__init__.py
@@ -1,7 +1,5 @@
-import logging
-logger = logging.getLogger(__name__)
-
-
+# noinspection PyUnusedLocal
 def includeme(config):
+    from magpie.common import get_logger
+    logger = get_logger(__name__)
     logger.info('Adding definitions...')
-