+++ magpie/definitions/alembic_definitions.py
@@ -6,3 +6,5 @@ from alembic import op                          # noqa: F401
 import alembic.config                           # noqa: F401
 # noinspection PyUnresolvedReferences
 import alembic.command                          # noqa: F401
+# noinspection PyUnresolvedReferences
+import alembic                                  # noqa: F401