+++ magpie/magpiectl.py
@@ -5,7 +5,7 @@
 Magpie is a service for AuthN and AuthZ based on Ziggurat-Foundations
 """
 
-from magpie.common import print_log, str2bool
+from magpie.common import print_log, str2bool, get_logger
 from magpie.constants import get_constant
 from magpie.helpers.register_default_users import register_default_users
 from magpie.register import (
@@ -15,10 +15,15 @@ from magpie.register import (
 from magpie.security import auth_config_from_settings
 from magpie import db, constants
 import os
+import sys
 # noinspection PyUnresolvedReferences
 import logging
-import logging.config   # find config in 'logging.ini'
-LOGGER = logging.getLogger(__name__)
+# import logging.config   # find config in 'logging.ini'
+LOGGER = get_logger(__name__)
+# transfer root logger handler to magpie to avoid duplicate logs
+if logging.root.handlers:
+    logging.root.handlers.pop(0)
+LOGGER.addHandler(logging.StreamHandler(sys.stderr))
 
 
 # noinspection PyUnusedLocal
@@ -29,13 +34,13 @@ def main(global_config=None, **settings):
     settings['magpie.root'] = constants.MAGPIE_ROOT
     settings['magpie.module'] = constants.MAGPIE_MODULE_DIR
 
-    print_log('Setting up loggers...')
+    print_log('Setting up loggers...', LOGGER)
     log_lvl = get_constant('MAGPIE_LOG_LEVEL', settings, 'magpie.log_level', default_value='INFO',
                            raise_missing=False, raise_not_set=False, print_missing=True)
     LOGGER.setLevel(log_lvl)
     sa_settings = db.set_sqlalchemy_log_level(log_lvl)
 
-    print_log('Looking for db migration requirement...')
+    print_log('Looking for db migration requirement...', LOGGER)
     db.run_database_migration_when_ready(settings)  # cannot pass db session as it might not even exist yet!
 
     # HACK:
@@ -48,10 +53,10 @@ def main(global_config=None, **settings):
     config_ini = get_constant('MAGPIE_INI_FILE_PATH', raise_missing=True)
     db_session = db.get_db_session_from_config_ini(config_ini, settings_override=sa_settings)
 
-    print_log('Register default users...')
+    print_log('Register default users...', LOGGER)
     register_default_users(db_session=db_session)
 
-    print_log('Register configuration providers...', LOGGER)
+    print_log('Register configuration providers...', logger=LOGGER)
     push_phoenix = str2bool(get_constant('PHOENIX_PUSH', settings=settings, settings_name='magpie.phoenix_push',
                                          raise_missing=False, raise_not_set=False, print_missing=True))
     prov_cfg = get_constant('MAGPIE_PROVIDERS_CONFIG_PATH', default_value='',
@@ -70,7 +75,7 @@ def main(global_config=None, **settings):
     else:
         print_log('No configuration file found for permissions registration, skipping...', LOGGER, logging.WARN)
 
-    print_log('Running configurations setup...')
+    print_log('Running configurations setup...', LOGGER)
     magpie_url_template = 'http://{hostname}:{port}'
     port = get_constant('MAGPIE_PORT', settings=settings, settings_name='magpie.port')
     if port:
@@ -89,7 +94,7 @@ def main(global_config=None, **settings):
     # config.scan('magpie')
     config.set_default_permission(get_constant('MAGPIE_ADMIN_PERMISSION'))
 
-    print_log('Starting Magpie app...')
+    print_log('Starting Magpie app...', LOGGER)
     wsgi_app = config.make_wsgi_app()
     return wsgi_app