+++ magpie/ui/swagger/__init__.py
@@ -1,12 +1,12 @@
 from magpie.api import api_rest_schemas as s
 from magpie.definitions.pyramid_definitions import NO_PERMISSION_REQUIRED
 from magpie.ui.swagger.views import api_swagger
-import logging
-logger = logging.getLogger(__name__)
+from magpie.common import get_logger
+LOGGER = get_logger(__name__)
 
 
 def includeme(config):
-    logger.info('Adding swagger ...')
+    LOGGER.info('Adding swagger ...')
     config.add_route(**s.service_api_route_info(s.SwaggerAPI))
     config.add_route(**s.service_api_route_info(s.SwaggerGenerator))
     config.add_view(s.api_schema, route_name=s.SwaggerGenerator.name, request_method='GET',